Move autocomplete actions to its own file and clears autocomplete slice of state if unneeded
